requirements.txt
google-generativeai
PyPDF2
python-docx
Pillow
python-dotenv